app/requirements.txt
streamlit
snowflake-connector-python
pandas
cd
python-dotenv
